refactor(ai/usage): one formatMoney in usage-poll for the Ink dashboard and accounts show, instead of two copies of the exponent rule
